Rename class `KeyboardButtonRequestUser` to `KeyboardButtonRequestUsers`, add new field
