+++ src/types.rs
@@ -131,7 +131,7 @@ pub mod invoice;
 pub mod keyboard_button;
 pub mod keyboard_button_poll_type;
 pub mod keyboard_button_request_chat;
-pub mod keyboard_button_request_user;
+pub mod keyboard_button_request_users;
 pub mod labeled_price;
 pub mod link_preview_options;
 pub mod location;
@@ -322,7 +322,7 @@ pub use invoice::Invoice;
 pub use keyboard_button::KeyboardButton;
 pub use keyboard_button_poll_type::KeyboardButtonPollType;
 pub use keyboard_button_request_chat::KeyboardButtonRequestChat;
-pub use keyboard_button_request_user::KeyboardButtonRequestUser;
+pub use keyboard_button_request_users::KeyboardButtonRequestUsers;
 pub use labeled_price::LabeledPrice;
 pub use link_preview_options::LinkPreviewOptions;
 pub use location::Location;
