+++ src/types/keyboard_button.rs
@@ -1,5 +1,5 @@
 use super::{
-    KeyboardButtonPollType, KeyboardButtonRequestChat, KeyboardButtonRequestUser, WebAppInfo,
+    KeyboardButtonPollType, KeyboardButtonRequestChat, KeyboardButtonRequestUsers, WebAppInfo,
 };
 
 use serde::{Deserialize, Serialize};
@@ -17,8 +17,8 @@ use serde_with::skip_serializing_none;
 pub struct KeyboardButton {
     /// Text of the button. If none of the optional fields are used, it will be sent as a message when the button is pressed
     pub text: String,
-    /// If specified, pressing the button will open a list of suitable users. Tapping on any user will send their identifier to the bot in a `user_shared` service message. Available in private chats only.
-    pub request_user: Option<KeyboardButtonRequestUser>,
+    /// If specified, pressing the button will open a list of suitable users. Identifiers of selected users will be sent to the bot in a `users_shared` service message. Available in private chats only.
+    pub request_users: Option<KeyboardButtonRequestUsers>,
     /// If specified, pressing the button will open a list of suitable chats. Tapping on a chat will send its identifier to the bot in a `chat_shared` service message. Available in private chats only.
     pub request_chat: Option<KeyboardButtonRequestChat>,
     /// If `true`, the user's phone number will be sent as a contact when the button is pressed. Available in private chats only.
@@ -36,7 +36,7 @@ impl KeyboardButton {
     pub fn new(text: impl Into<String>) -> Self {
         Self {
             text: text.into(),
-            request_user: None,
+            request_users: None,
             request_chat: None,
             request_contact: None,
             request_location: None,
@@ -54,9 +54,9 @@ impl KeyboardButton {
     }
 
     #[must_use]
-    pub fn request_user(self, val: KeyboardButtonRequestUser) -> Self {
+    pub fn request_user(self, val: KeyboardButtonRequestUsers) -> Self {
         Self {
-            request_user: Some(val),
+            request_users: Some(val),
             ..self
         }
     }
@@ -104,9 +104,9 @@ impl KeyboardButton {
 
 impl KeyboardButton {
     #[must_use]
-    pub fn request_user_option(self, val: Option<KeyboardButtonRequestUser>) -> Self {
+    pub fn request_user_option(self, val: Option<KeyboardButtonRequestUsers>) -> Self {
         Self {
-            request_user: val,
+            request_users: val,
             ..self
         }
     }
